% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conectar_bd.R
\name{conectar_bd}
\alias{conectar_bd}
\title{Conectar a la base de datos}
\usage{
conectar_bd(archivo_credenciales = "credenciales.R")
}
\arguments{
\item{archivo_credenciales}{Un archivo de texto R que contiene las credenciales
necesarias para la conexi贸n a la base de datos. El archivo debe definir las
variables `db_host`, `db_port`, `db_name`, `db_user`, y `db_password`. El
valor por defecto es `"credenciales.R"`.}
}
\value{
Un objeto de conexi贸n de base de datos que puede ser utilizado con las
funciones del paquete DBI para interactuar con la base de datos.
}
\description{
Esta funci贸n establece una conexi贸n con la base de datos MySQL utilizando las
credenciales cargadas desde un archivo R externo.
}
\examples{
# Conectar a la base de datos utilizando las credenciales predeterminadas
con <- conectar_bd("credenciales.R")

}
