% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{crear_df_vacio}
\alias{crear_df_vacio}
\title{Crear Dataframe Vacio Estandarizado}
\usage{
crear_df_vacio()
}
\value{
Un data frame con 0 filas y las columnas estandar.
}
\description{
Crea un dataframe vacio con la estructura estandar del paquete datamedios.
Asegura consistencia en las columnas y tipos de datos entre todos los scrapers.
}
\keyword{internal}
