% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desconectar_bd.R
\name{desconectar_bd}
\alias{desconectar_bd}
\title{Desconectar la base de datos}
\usage{
desconectar_bd(con)
}
\arguments{
\item{con}{Un objeto de conexión creado por la función `conectar_bd()`.}
}
\value{
No retorna nada. Solo cierra la conexión.
}
\description{
Esta función cierra la conexión activa a la base de datos.
}
\examples{
con <- conectar_bd("credenciales.R")
desconectar_bd(con)

}
