% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraccion_parrafos.R
\name{extracción_parrafos}
\alias{extracción_parrafos}
\title{Extraer párrafos de una columna de texto}
\usage{
extracción_parrafos(datos, sinonimos)
}
\arguments{
\item{datos}{Data frame que contiene los datos de entrada con la columna "post_content".}

\item{sinonimos}{Vector de sinónimos que se incluirán en la búsqueda.}
}
\value{
Data frame con una columna adicional 'parrafos_filtrados' que contiene los párrafos extraídos como listas.
}
\description{
Esta función procesa una columna de texto en un dataframe y extrae los párrafos que coinciden con los sinónimos proporcionados.
}
\examples{
datos <- extracción_párrafos(datos, sinonimos = c("IA", "AI"))
print(datos$parrafos_filtrados[[1]])

}
