% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_fecha_ciper.R
\name{extraer_noticias_fecha_ciper}
\alias{extraer_noticias_fecha_ciper}
\title{Extrae noticias de Ciper Chile por rango de fechas}
\usage{
extraer_noticias_fecha_ciper(search_query = NULL, fecha_inicio, fecha_fin)
}
\arguments{
\item{search_query}{El termino de busqueda (opcional).}

\item{fecha_inicio}{Fecha de inicio del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{fecha_fin}{Fecha de fin del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}
}
\value{
Un dataframe con las noticias extraidas, estandarizado al formato de datamedios.
}
\description{
Esta funcion se conecta a la API de Ciper para descargar noticias dentro de un rango de fechas.
}
