% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_fecha_emol.R
\name{extraer_noticias_fecha_emol}
\alias{extraer_noticias_fecha_emol}
\title{Extraccion de noticias de emol.com por rango de fechas}
\usage{
extraer_noticias_fecha_emol(search_query, fecha_inicio, fecha_fin, fuente)
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{fecha_inicio}{Fecha de inicio del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{fecha_fin}{Fecha de fin del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{fuente}{Fuente de emol para iterar (obligatoria).}
}
\value{
Un dataframe con las noticias extraidas.
}
\description{
Esta funcion permite realizar una extraccion automatizada de noticias de emol.com utilizando un rango de fechas.
}
\examples{
\dontrun{
noticias <- extraer_noticias_fecha_emol("inteligencia artificial", "2025-01-01",
  "2025-02-24",
  fuente = "emol"
)
}
}
