% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_max_res_ciper.R
\name{extraer_noticias_max_res_ciper}
\alias{extraer_noticias_max_res_ciper}
\title{Extrae noticias de Ciper Chile}
\usage{
extraer_noticias_max_res_ciper(search_query = NULL, max_results = NULL)
}
\arguments{
\item{search_query}{El termino de busqueda (opcional). Si es NULL, extrae todos los articulos.}

\item{max_results}{El numero maximo de noticias a extraer.}
}
\value{
Un dataframe con las noticias extraidas, estandarizado al formato de datamedios.
}
\description{
Esta funcion se conecta a la API de Ciper para descargar noticias.
}
\examples{
\donttest{
# Extraer los ultimos 5 articulos con una busqueda
noticias_ciper <- extraer_noticias_max_res_ciper("corrupcion", max_results = 5)

# Extraer los ultimos 10 articulos sin busqueda
ultimos_ciper <- extraer_noticias_max_res_ciper(max_results = 10)
}
}
