% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_max_res_emol.R
\name{extraer_noticias_max_res_emol}
\alias{extraer_noticias_max_res_emol}
\title{Extraccion de noticias de Emol.com}
\usage{
extraer_noticias_max_res_emol(search_query, max_results = NULL, fuente)
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{max_results}{Numero maximo de resultados a extraer (opcional, por defecto todos).}

\item{fuente}{Fuente de emol para iterar (obligatoria).}
}
\value{
Un dataframe con las noticias extraidas.
}
\description{
Esta funcion permite extraer noticias de las fuentes de Emol, tanto de las
noticias no pagas de emol, como de guioteca y los medios regionales de El Mercurio
}
