% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_fuentes_emol}
\alias{get_fuentes_emol}
\title{Obtener Fuentes de Emol}
\usage{
get_fuentes_emol()
}
\value{
Vector de caracteres con las fuentes.
}
\description{
Devuelve un vector con los nombres de las fuentes disponibles en el grupo Emol.
Util para iterar sobre todas las fuentes cuando se solicita 'emol-todas'.
}
\keyword{internal}
