% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_random_user_agent}
\alias{get_random_user_agent}
\title{Obtener User-Agent Aleatorio}
\usage{
get_random_user_agent()
}
\value{
Cadena de caracteres con un User-Agent.
}
\description{
Devuelve una cadena de User-Agent seleccionada aleatoriamente de una lista predefinida.
Util para rotar identificadores y evitar bloqueos simples en scrapers.
}
\keyword{internal}
