% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_req_emol.R
\name{init_req_emol}
\alias{init_req_emol}
\title{Inicializa una solicitud a emol.com y retorna el primer caso de busqueda}
\usage{
init_req_emol(search_query, fuentes = "emol-todas")
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{fuentes}{Un string donde se ponen las fuentes de emol a consultar}
}
\value{
Un dataframe con el primer caso de la busqueda.
}
\description{
Esta funcion permite realizar una consulta inicial a emol.com utilizando una frase de busqueda.
}
\examples{
\dontrun{
primer_caso <- init_req_emol("Boric", fuentes="emol")
}
}
