% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteracion_emol.R
\name{iteracion_emol}
\alias{iteracion_emol}
\title{Inicializa una solicitud a emol.com y retorna maximo 10 noticias}
\usage{
iteracion_emol(search_query, page = 0, fuentes = "emol-todas")
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{page}{La pagina de busqueda para iterar, es un int}

\item{fuentes}{Es un string que deberia tener solo fuentes de emol posibles separadas por comas.}
}
\value{
Un dataframe con el caso de la busqueda, incluyendo solo columnas especificas.
}
\description{
Esta funcion auxiliar llama a emol.com utilizando una frase de busqueda.
Entrega como maximo 10 resultados. Se debe llamar desde otras funciones solo
con una fuente a la vez, es decir, sin llamar a emol-todas.
}
\examples{
\dontrun{
primer_caso <- iteracion_emol("Boric", fuentes="emol-todas")
}
}
