% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpieza_notas.R
\name{limpieza_notas}
\alias{limpieza_notas}
\title{Función para limpiar notas de contenido HTML}
\usage{
limpieza_notas(datos, sinonimos = c())
}
\arguments{
\item{datos}{Data frame donde están almacenadas las notas.}

\item{sinonimos}{Una lista}
}
\value{
Un listado de character que sirva para agregar sinónimos del search_query del data frame
}
\description{
Esta función permite limpiar por completo las notas eliminando códigos y secciones irrelevantes.
Verifica que el input sea un data frame con una columna llamada `post_content`.
}
\examples{
limpieza_notas(datos, sinonimos = c("IA", "AI"))
}
