% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserFuentes.R
\name{parserFuentes}
\alias{parserFuentes}
\title{Parser de Fuentes}
\usage{
parserFuentes(cadena)
}
\arguments{
\item{cadena}{Un string que contiene nombres de fuentes separados por comas.}
}
\value{
Una lista de strings, cada uno representando una fuente sin espacios en blanco adicionales.
}
\description{
Esta funcion toma un string que contiene nombres de fuentes separados por comas
y devuelve una lista con cada fuente como un elemento separado, sin espacios en blanco adicionales.
}
\examples{
parserFuentes("bbcl, emol, mediosregionales, ")
parserFuentes(" emol-todas, bbcl")
}
