% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bbcl-helpers.R
\name{procesar_data_bbcl}
\alias{procesar_data_bbcl}
\title{Procesar y estandarizar datos de BBCL (Helper)}
\usage{
procesar_data_bbcl(all_data, search_query)
}
\arguments{
\item{all_data}{Data frame con los datos crudos extraidos de la API.}

\item{search_query}{La frase de busqueda utilizada.}
}
\value{
Un data frame estandarizado.
}
\description{
Esta funcion centraliza la logica de limpieza y estandarizacion de datos crudos
provenientes de la API de BioBioChile, convirtiendolos al formato estandar
del paquete datamedios.
}
