% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_body_ciper.R
\name{scrape_body_ciper}
\alias{scrape_body_ciper}
\title{Scrapea el cuerpo de las noticias desde una lista de URLs de Ciper}
\usage{
scrape_body_ciper(news_df, parallel = TRUE)
}
\arguments{
\item{news_df}{Un dataframe que debe contener una columna ''url''.}

\item{parallel}{Booleano que indica si se debe usar procesamiento en paralelo. Por defecto es TRUE.}
}
\value{
El dataframe de entrada con la columna ''contenido_limpio'' rellenada con el texto scrapeado.
}
\description{
Esta funcion itera sobre un dataframe de noticias, visita la URL de cada una,
y extrae el texto principal usando un selector CSS especifico para Ciper.
Admite ejecucion en paralelo para acelerar el proceso.
}
\keyword{internal}
