% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacion.R
\name{visualizacion}
\alias{visualizacion}
\title{Función de visualizaciones}
\usage{
visualizacion(datos, tipo, stop_words, ...)
}
\arguments{
\item{datos}{data frame donde se desea aplicar los gráficos.}

\item{tipo}{Tipo de visualización que se desea (word_cloud, time_frecuency, top_words, tf-idf, heatmap).}

\item{stop_words}{Definir las palabras que serán ignoradas en la visualización. Puede ser es (realizado por los desarrolladores del paquete), null, o una lista.}

\item{...}{Parámetros adicionales necesarios para ciertos tipos de visualización (e.g., para word_cloud).}
}
\value{
Un gráfico con características dependientes del tipo especificado y los datos ingresados.
}
\description{
Esta función incluye varias formas de visualizar los datos obtenidos.
}
\examples{
datos <- extraer_noticias_fecha("cambio climatico", "2024-01-01","2025-01-01", subir_a_bd = FALSE)
datos <- extraccion_parrafos(datos)
datos_proc <- limpieza_notas(datos)
visualizacion(datos_proc, word_cloud, stop_words=es)
}
