% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_datos_unicos.R
\name{agregar_datos_unicos}
\alias{agregar_datos_unicos}
\title{Agregar datos únicos a una tabla MySQL}
\usage{
agregar_datos_unicos(data)
}
\arguments{
\item{data}{Un data frame con los datos a insertar.}
}
\description{
Esta función agrega datos a una tabla MySQL utilizando una API que espera datos en formato de lista.
}
\examples{
\dontrun{
# Agregar datos únicos
agregar_datos_unicos(noticias)
}

}
