% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_datos_unicos.R
\name{agregar_datos_unicos}
\alias{agregar_datos_unicos}
\title{Agregar datos unicos a una tabla MySQL}
\usage{
agregar_datos_unicos(data)
}
\arguments{
\item{data}{Un data frame con los datos a insertar.}
}
\description{
Esta funcion agrega datos a una tabla MySQL utilizando una API que espera datos en formato JSON.
}
\examples{
\dontrun{
# Agregar datos unicos
noticias <- extraer_noticias_max_res("tesla", max_results=10, subir_a_bd = FALSE)
agregar_datos_unicos(noticias)
}

}
