% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_fecha.R
\name{extraer_noticias_fecha}
\alias{extraer_noticias_fecha}
\title{Extracción de noticias desde la API de BíoBío.cl por rango de fechas}
\usage{
extraer_noticias_fecha(search_query, fecha_inicio, fecha_fin)
}
\arguments{
\item{search_query}{Una frase de búsqueda (obligatoria).}

\item{fecha_inicio}{Fecha de inicio del rango de búsqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{fecha_fin}{Fecha de fin del rango de búsqueda en formato "YYYY-MM-DD" (obligatoria).}
}
\value{
Un dataframe con las noticias extraídas.
}
\description{
Esta función permite realizar una extracción automatizada de noticias desde la API de BíoBío.cl utilizando un rango de fechas.
}
\examples{
noticias <- extraer_noticias_fecha("inteligencia artificial", "2023-01-01", "2023-12-31")
}
