% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_fecha.R
\name{extraer_noticias_fecha}
\alias{extraer_noticias_fecha}
\title{Extraccion de noticias desde la API de BioBio.cl por rango de fechas}
\usage{
extraer_noticias_fecha(
  search_query,
  fecha_inicio,
  fecha_fin,
  subir_a_bd = TRUE
)
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{fecha_inicio}{Fecha de inicio del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{fecha_fin}{Fecha de fin del rango de búsqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{subir_a_bd}{por defecto TRUE, FALSE para test y cosas por el estilo (opcional).}
}
\value{
Un dataframe con las noticias extraidas.
}
\description{
Esta funcion permite realizar una extraccion automatizada de noticias desde la API de BioBio.cl utilizando un rango de fechas.
}
\examples{
noticias <- extraer_noticias_fecha("inteligencia artificial", "2023-01-01", "2023-12-31", subir_a_bd = FALSE)
}
