% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_fecha_bbcl.R
\name{extraer_noticias_fecha_bbcl}
\alias{extraer_noticias_fecha_bbcl}
\title{Extraccion de noticias de BioBio.cl por rango de fechas}
\usage{
extraer_noticias_fecha_bbcl(search_query, fecha_inicio, fecha_fin)
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{fecha_inicio}{Fecha de inicio del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}

\item{fecha_fin}{Fecha de fin del rango de busqueda en formato "YYYY-MM-DD" (obligatoria).}
}
\value{
Un dataframe con las noticias extraidas.
}
\description{
Esta funcion permite realizar una extraccion automatizada de noticias de BioBio.cl utilizando un rango de fechas.
}
\examples{
noticias <- extraer_noticias_fecha_bbcl("inteligencia artificial", "2025-01-01",
"2025-02-24")
}
