% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_max_res.R
\name{extraer_noticias_max_res}
\alias{extraer_noticias_max_res}
\title{Extracción de noticias desde la API de BíoBío.cl}
\usage{
extraer_noticias_max_res(search_query, max_results = NULL)
}
\arguments{
\item{search_query}{Una frase de búsqueda (obligatoria).}

\item{max_results}{Número máximo de resultados a extraer (opcional, por defecto todos).}
}
\value{
Un dataframe con las noticias extraídas.
}
\description{
Esta función permite realizar una extracción automatizada de noticias desde la API de BíoBío.cl.
}
\examples{
noticias <- extraer_noticias_max_res("inteligencia artificial", max_results = 100)
}
