% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_noticias_max_res_bbcl.R
\name{extraer_noticias_max_res_bbcl}
\alias{extraer_noticias_max_res_bbcl}
\title{Extraccion de noticias de BioBio.cl por cantidad maxima de resultados}
\usage{
extraer_noticias_max_res_bbcl(search_query, max_results = NULL)
}
\arguments{
\item{search_query}{Una frase de busqueda (obligatoria).}

\item{max_results}{Cantidad maxima de resultados (opcional).}
}
\value{
Un dataframe con las noticias extraidas.
}
\description{
Esta funcion permite realizar una extraccion automatizada de noticias de BioBio.cl
entregando como parametro una cantidad maxima de resultados.
}
\examples{
\dontrun{
noticias <- extraer_noticias_fecha_bbcl("inteligencia artificial", "2025-01-01",
"2025-02-24")
}
}
