% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_notas_fecha.R
\name{grafico_notas_fecha}
\alias{grafico_notas_fecha}
\title{Grafico de notas por periodo}
\usage{
grafico_notas_fecha(
  datos,
  titulo,
  fecha_inicio = NULL,
  fecha_fin = NULL,
  agrupar_por = "day"
)
}
\arguments{
\item{datos}{Data frame con los datos procesados, que debe incluir la columna `fecha` en formato YYYY-MM-DD.}

\item{titulo}{Texto que aparecera en el titulo del grafico.}

\item{fecha_inicio}{Fecha de inicio para la construccion del grafico en formato YYYY-MM-DD (opcional).}

\item{fecha_fin}{Fecha de finalizacion para la construccion del grafico en formato YYYY-MM-DD (opcional).}

\item{agrupar_por}{Cadena de texto que especifica el periodo de agrupacion.
Valores validos son `"day"` (por defecto) o `"month"`.}
}
\value{
Un grafico ggplot2 que muestra la cantidad de publicaciones por el periodo seleccionado.
}
\description{
Esta funcion genera un grafico que muestra la cantidad de publicaciones
agrupadas por dia o por mes.
}
\examples{
\dontrun{
# Ejemplo original, agrupando por mes
datos <- extraer_noticias_fecha("cambio climatico", "2024-01-01","2025-01-01", subir_a_bd = FALSE)
grafico_notas_por_mes(datos, titulo = "Cambio Climatico (por mes)",
                      agrupar_por = "month")
# Devuelve un grafico de barras con la cantidad de notas por mes

# Nuevo ejemplo, agrupando por dia para mayor granularidad
grafico_notas_por_mes(datos, titulo = "Cambio Climatico (por dia)",
                      fecha_inicio = "2024-01-01", fecha_fin = "2024-03-31")
# Devuelve un grafico de barras con la cantidad de notas por dia
}

}
