% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_notas_por_mes.R
\name{grafico_notas_por_mes}
\alias{grafico_notas_por_mes}
\title{Gráfico de notas por mes}
\usage{
grafico_notas_por_mes(datos, titulo, fecha_inicio = NULL, fecha_fin = NULL)
}
\arguments{
\item{datos}{Data frame con los datos procesados, que debe incluir la columna `raw_post_date` en formato YYYY-MM-DD.}

\item{titulo}{Texto que aparecerá en el título del gráfico.}

\item{fecha_inicio}{Fecha de inicio para la construcción del gráfico en formato YYYY-MM-DD (opcional).}

\item{fecha_fin}{Fecha de finalización para la construcción del gráfico en formato YYYY-MM-DD (opcional).}
}
\value{
Un gráfico ggplot2 que muestra la cantidad de publicaciones por mes.
}
\description{
Esta función genera un gráfico de línea que muestra la cantidad de publicaciones agrupadas por mes.
}
\examples{
grafico_notas_por_mes(datos_proc, titulo = "Cambio Climático", fecha_inicio = "2024-01-01", fecha_fin = "2025-01-01")
grafico_notas_por_mes(datos_proc, titulo = "Cambio Climático")
}
