% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_req_bbcl.R
\name{init_req_bbcl}
\alias{init_req_bbcl}
\title{Inicializa una solicitud a la API de BíoBío.cl y retorna el primer caso de búsqueda}
\usage{
init_req_bbcl(search_query)
}
\arguments{
\item{search_query}{Una frase de búsqueda (obligatoria).}
}
\value{
Un dataframe con el primer caso de la búsqueda.
}
\description{
Esta función permite realizar una consulta inicial a la API de BíoBío.cl utilizando una frase de búsqueda.
}
\examples{
primer_caso <- init_req_bbcl("inteligencia artificial")
}
