% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpieza_notas.R
\name{limpieza_notas}
\alias{limpieza_notas}
\title{Funcion para limpiar notas de contenido HTML}
\usage{
limpieza_notas(datos, sinonimos = c())
}
\arguments{
\item{datos}{Data frame donde estan almacenadas las notas y con la funcion extraccion_parrafos ya operada.}

\item{sinonimos}{Una lista de character}
}
\value{
Un dataframe con el contenido limpio en la columna contenido_limpio
}
\description{
Esta funcion permite limpiar por completo las notas eliminando codigos y secciones irrelevantes.
Verifica que el input sea un data frame con una columna llamada `contenido`.
}
\examples

\dontrun{
datos <- extraer_noticias_max_res("inteligencia artificial", max_results= 150, subir_a_bd = FALSE)
datos <- extraccion_parrafos(datos)
datos_proc <- limpieza_notas(datos, sinonimos = c("IA", "AI")
}

