% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud.R
\name{word_cloud}
\alias{word_cloud}
\title{Función de visualizaciones}
\usage{
word_cloud(datos, max_words, stop_words = NULL)
}
\arguments{
\item{datos}{data frame que genera la función de extracción, con el texto ya limpio.}

\item{max_words}{Cantidad máxima de palabras que aparecerán en la nube.}

\item{stop_words}{Definir las palabras que serán ignoradas en la visualización. Puede ser 'es' (realizado por los desarrolladores del paquete), NULL, o una lista.}
}
\value{
Una nube de palabras con las palabras más frecuentes.
}
\description{
Esta función incluye varias formas de visualizar los datos obtenidos.
}
\examples{
visualizacion(datos_proc, max_words = 50, stop_words = es)
}
