% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud.R
\name{word_cloud}
\alias{word_cloud}
\title{Funcion de visualizaciones}
\usage{
word_cloud(datos, max_words, stop_words = NULL)
}
\arguments{
\item{datos}{data frame que genera la funcion de extraccion, con el texto ya limpio.}

\item{max_words}{Cantidad maxima de palabras que apareceran en la nube.}

\item{stop_words}{Definir las palabras que seran ignoradas en la visualizacion. Puede ser 'es' (realizado por los desarrolladores del paquete), NULL, o una lista.}
}
\value{
Una nube de palabras con las palabras mas frecuentes.
}
\description{
Esta funcion incluye varias formas de visualizar los datos obtenidos.
}
\examples{
datos <- extraer_noticias_fecha("cambio climatico", "2024-01-01","2025-01-01", subir_a_bd = FALSE)
datos <- extraccion_parrafos(datos)
datos_proc <- limpieza_notas(datos)
word_cloud(datos_proc, max_words = 50, stop_words = es)
}
